inline everything into create_artifacts_from_tracer
