+++ src/witness/postprocessing/mod.rs
@@ -91,495 +91,6 @@ use crate::boojum::field::SmallField;
 use crate::boojum::gadgets::traits::allocatable::CSAllocatableExt;
 use crate::boojum::gadgets::traits::round_function::*;
 
-pub fn create_leaf_level_circuits_and_scheduler_witness<
-    CB: FnMut(
-        ZkSyncBaseLayerCircuit<
-            GoldilocksField,
-            VmWitnessOracle<GoldilocksField>,
-            Poseidon2Goldilocks,
-        >,
-    ),
-    QSCB: FnMut(
-        u8,
-        RecursionQueueSimulator<GoldilocksField>,
-        Vec<ClosedFormInputCompactFormWitness<GoldilocksField>>,
-    ),
->(
-    zkporter_is_available: bool,
-    default_aa_code_hash: U256,
-    vm_instances_witness: Vec<VmInstanceWitness<GoldilocksField, VmWitnessOracle<GoldilocksField>>>,
-    artifacts: FullBlockArtifacts<GoldilocksField>,
-    geometry: GeometryConfig,
-    round_function: &Poseidon2Goldilocks,
-    mut circuit_callback: CB,
-    mut queue_simulator_callback: QSCB,
-) -> (
-    BlockFirstAndLastBasicCircuits,
-    Vec<ClosedFormInputCompactFormWitness<GoldilocksField>>,
-) {
-    assert!(artifacts.is_processed);
-
-    let FullBlockArtifacts {
-        ram_permutation_circuits_data,
-        code_decommitter_circuits_data,
-        decommittments_deduplicator_circuits_data,
-        log_demuxer_circuit_data,
-        storage_deduplicator_circuit_data,
-        events_deduplicator_circuit_data,
-        l1_messages_deduplicator_circuit_data,
-        rollup_storage_application_circuit_data,
-        keccak256_circuits_data,
-        sha256_circuits_data,
-        ecrecover_circuits_data,
-        l1_messages_linear_hash_data,
-        ..
-    } = artifacts;
-
-    let round_function = Arc::new(round_function.clone());
-
-    use crate::zkevm_circuits::base_structures::vm_state::GlobalContextWitness;
-
-    let in_circuit_global_context = GlobalContextWitness {
-        zkporter_is_available,
-        default_aa_code_hash,
-    };
-
-    use crate::witness::utils::create_cs_for_witness_generation;
-    use crate::witness::utils::simulate_public_input_value_from_witness;
-
-    let mut cs_for_witness_generation =
-        create_cs_for_witness_generation::<GoldilocksField, Poseidon2Goldilocks>(
-            TRACE_LEN_LOG_2_FOR_CALCULATION,
-            MAX_VARS_LOG_2_FOR_CALCULATION,
-        );
-    let mut cycles_used: usize = 0;
-
-    let mut all_compact_forms = vec![];
-
-    // VM
-
-    let mut main_vm_circuits = FirstAndLastCircuit::default();
-    let mut main_vm_circuits_compact_forms_witnesses = vec![];
-    let mut queue_simulator = RecursionQueueSimulator::empty();
-    let num_instances = vm_instances_witness.len();
-    let mut observable_input = None;
-    for (instance_idx, vm_instance) in vm_instances_witness.into_iter().enumerate() {
-        use crate::witness::utils::vm_instance_witness_to_circuit_formal_input;
-        let is_first = instance_idx == 0;
-        let is_last = instance_idx == num_instances - 1;
-        let mut circuit_input = vm_instance_witness_to_circuit_formal_input(
-            vm_instance,
-            is_first,
-            is_last,
-            in_circuit_global_context.clone(),
-        );
-
-        if observable_input.is_none() {
-            assert!(is_first);
-            observable_input = Some(circuit_input.closed_form_input.observable_input.clone());
-        } else {
-            circuit_input.closed_form_input.observable_input =
-                observable_input.as_ref().unwrap().clone();
-        }
-
-        let (proof_system_input, compact_form_witness) = simulate_public_input_value_from_witness(
-            &mut cs_for_witness_generation,
-            circuit_input.closed_form_input.clone(),
-            &*round_function,
-        );
-
-        cycles_used += 1;
-        if cycles_used == CYCLES_PER_SCRATCH_SPACE {
-            cs_for_witness_generation =
-                create_cs_for_witness_generation::<GoldilocksField, Poseidon2Goldilocks>(
-                    TRACE_LEN_LOG_2_FOR_CALCULATION,
-                    MAX_VARS_LOG_2_FOR_CALCULATION,
-                );
-            cycles_used = 0;
-        }
-
-        let instance = VMMainCircuit {
-            witness: AtomicCell::new(Some(circuit_input)),
-            config: Arc::new(geometry.cycles_per_vm_snapshot as usize),
-            round_function: round_function.clone(),
-            expected_public_input: Some(proof_system_input),
-        };
-
-        if is_first {
-            main_vm_circuits.first = Some(instance.clone());
-        }
-        if is_last {
-            main_vm_circuits.last = Some(instance.clone());
-        }
-
-        let instance = ZkSyncBaseLayerCircuit::MainVM(instance);
-
-        let recursive_request = RecursionRequest {
-            circuit_type: GoldilocksField::from_u64_unchecked(
-                instance.numeric_circuit_type() as u64
-            ),
-            public_input: proof_system_input,
-        };
-        let _ = queue_simulator.push(recursive_request, &*round_function);
-
-        circuit_callback(instance);
-        main_vm_circuits_compact_forms_witnesses.push(compact_form_witness);
-    }
-    all_compact_forms.extend(main_vm_circuits_compact_forms_witnesses.clone());
-    queue_simulator_callback(
-        BaseLayerCircuitType::VM as u8,
-        queue_simulator,
-        main_vm_circuits_compact_forms_witnesses,
-    );
-
-    // Code decommitter sorter
-    let circuit_type = BaseLayerCircuitType::DecommitmentsFilter;
-
-    let mut maker = CircuitMaker::new(
-        geometry.cycles_code_decommitter_sorter,
-        round_function.clone(),
-        &mut cs_for_witness_generation,
-        &mut cycles_used,
-    );
-
-    for circuit_input in decommittments_deduplicator_circuits_data.into_iter() {
-        circuit_callback(ZkSyncBaseLayerCircuit::CodeDecommittmentsSorter(
-            maker.process(circuit_input, circuit_type),
-        ));
-    }
-
-    let (
-        code_decommittments_sorter_circuits,
-        queue_simulator,
-        code_decommittments_sorter_circuits_compact_forms_witnesses,
-    ) = maker.into_results();
-    all_compact_forms.extend(code_decommittments_sorter_circuits_compact_forms_witnesses.clone());
-    queue_simulator_callback(
-        circuit_type as u8,
-        queue_simulator,
-        code_decommittments_sorter_circuits_compact_forms_witnesses,
-    );
-
-    // Actual decommitter
-    let circuit_type = BaseLayerCircuitType::Decommiter;
-
-    let mut maker = CircuitMaker::new(
-        geometry.cycles_per_code_decommitter,
-        round_function.clone(),
-        &mut cs_for_witness_generation,
-        &mut cycles_used,
-    );
-
-    for circuit_input in code_decommitter_circuits_data.into_iter() {
-        circuit_callback(ZkSyncBaseLayerCircuit::CodeDecommitter(
-            maker.process(circuit_input, circuit_type),
-        ));
-    }
-
-    let (
-        code_decommitter_circuits,
-        queue_simulator,
-        code_decommitter_circuits_compact_forms_witnesses,
-    ) = maker.into_results();
-    all_compact_forms.extend(code_decommitter_circuits_compact_forms_witnesses.clone());
-    queue_simulator_callback(
-        circuit_type as u8,
-        queue_simulator,
-        code_decommitter_circuits_compact_forms_witnesses,
-    );
-
-    // log demux
-    let circuit_type = BaseLayerCircuitType::LogDemultiplexer;
-
-    let mut maker = CircuitMaker::new(
-        geometry.cycles_per_log_demuxer,
-        round_function.clone(),
-        &mut cs_for_witness_generation,
-        &mut cycles_used,
-    );
-
-    for circuit_input in log_demuxer_circuit_data.into_iter() {
-        circuit_callback(ZkSyncBaseLayerCircuit::LogDemuxer(
-            maker.process(circuit_input, circuit_type),
-        ));
-    }
-
-    let (log_demux_circuits, queue_simulator, log_demux_circuits_compact_forms_witnesses) =
-        maker.into_results();
-    all_compact_forms.extend(log_demux_circuits_compact_forms_witnesses.clone());
-    queue_simulator_callback(
-        circuit_type as u8,
-        queue_simulator,
-        log_demux_circuits_compact_forms_witnesses,
-    );
-
-    // keccak precompiles
-    let circuit_type = BaseLayerCircuitType::KeccakPrecompile;
-
-    let mut maker = CircuitMaker::new(
-        geometry.cycles_per_keccak256_circuit,
-        round_function.clone(),
-        &mut cs_for_witness_generation,
-        &mut cycles_used,
-    );
-
-    for circuit_input in keccak256_circuits_data.into_iter() {
-        circuit_callback(ZkSyncBaseLayerCircuit::KeccakRoundFunction(
-            maker.process(circuit_input, circuit_type),
-        ));
-    }
-
-    let (
-        keccak_precompile_circuits,
-        queue_simulator,
-        keccak_precompile_circuits_compact_forms_witnesses,
-    ) = maker.into_results();
-    all_compact_forms.extend(keccak_precompile_circuits_compact_forms_witnesses.clone());
-    queue_simulator_callback(
-        circuit_type as u8,
-        queue_simulator,
-        keccak_precompile_circuits_compact_forms_witnesses,
-    );
-
-    // sha256 precompiles
-    let circuit_type = BaseLayerCircuitType::Sha256Precompile;
-
-    let mut maker = CircuitMaker::new(
-        geometry.cycles_per_sha256_circuit,
-        round_function.clone(),
-        &mut cs_for_witness_generation,
-        &mut cycles_used,
-    );
-
-    for circuit_input in sha256_circuits_data.into_iter() {
-        circuit_callback(ZkSyncBaseLayerCircuit::Sha256RoundFunction(
-            maker.process(circuit_input, circuit_type),
-        ));
-    }
-
-    let (
-        sha256_precompile_circuits,
-        queue_simulator,
-        sha256_precompile_circuits_compact_forms_witnesses,
-    ) = maker.into_results();
-    all_compact_forms.extend(sha256_precompile_circuits_compact_forms_witnesses.clone());
-    queue_simulator_callback(
-        circuit_type as u8,
-        queue_simulator,
-        sha256_precompile_circuits_compact_forms_witnesses,
-    );
-
-    // ecrecover precompiles
-    let circuit_type = BaseLayerCircuitType::EcrecoverPrecompile;
-
-    let mut maker = CircuitMaker::new(
-        geometry.cycles_per_ecrecover_circuit,
-        round_function.clone(),
-        &mut cs_for_witness_generation,
-        &mut cycles_used,
-    );
-
-    for circuit_input in ecrecover_circuits_data.into_iter() {
-        circuit_callback(ZkSyncBaseLayerCircuit::ECRecover(
-            maker.process(circuit_input, circuit_type),
-        ));
-    }
-
-    let (
-        ecrecover_precompile_circuits,
-        queue_simulator,
-        ecrecover_precompile_circuits_compact_forms_witnesses,
-    ) = maker.into_results();
-    all_compact_forms.extend(ecrecover_precompile_circuits_compact_forms_witnesses.clone());
-    queue_simulator_callback(
-        circuit_type as u8,
-        queue_simulator,
-        ecrecover_precompile_circuits_compact_forms_witnesses,
-    );
-
-    // RAM permutation
-    let circuit_type = BaseLayerCircuitType::RamValidation;
-
-    let mut maker = CircuitMaker::new(
-        geometry.cycles_per_ram_permutation,
-        round_function.clone(),
-        &mut cs_for_witness_generation,
-        &mut cycles_used,
-    );
-
-    for circuit_input in ram_permutation_circuits_data.into_iter() {
-        circuit_callback(ZkSyncBaseLayerCircuit::RAMPermutation(
-            maker.process(circuit_input, circuit_type),
-        ));
-    }
-
-    let (
-        ram_permutation_circuits,
-        queue_simulator,
-        ram_permutation_circuits_compact_forms_witnesses,
-    ) = maker.into_results();
-    all_compact_forms.extend(ram_permutation_circuits_compact_forms_witnesses.clone());
-    queue_simulator_callback(
-        circuit_type as u8,
-        queue_simulator,
-        ram_permutation_circuits_compact_forms_witnesses,
-    );
-
-    // storage sorter
-    let circuit_type = BaseLayerCircuitType::StorageFilter;
-
-    let mut maker = CircuitMaker::new(
-        geometry.cycles_per_storage_sorter,
-        round_function.clone(),
-        &mut cs_for_witness_generation,
-        &mut cycles_used,
-    );
-
-    for circuit_input in storage_deduplicator_circuit_data.into_iter() {
-        circuit_callback(ZkSyncBaseLayerCircuit::StorageSorter(
-            maker.process(circuit_input, circuit_type),
-        ));
-    }
-
-    let (storage_sorter_circuits, queue_simulator, storage_sorter_circuit_compact_form_witnesses) =
-        maker.into_results();
-    all_compact_forms.extend(storage_sorter_circuit_compact_form_witnesses.clone());
-    queue_simulator_callback(
-        circuit_type as u8,
-        queue_simulator,
-        storage_sorter_circuit_compact_form_witnesses,
-    );
-
-    // storage application
-    let circuit_type = BaseLayerCircuitType::StorageApplicator;
-
-    let mut maker = CircuitMaker::new(
-        geometry.cycles_per_storage_application,
-        round_function.clone(),
-        &mut cs_for_witness_generation,
-        &mut cycles_used,
-    );
-
-    for circuit_input in rollup_storage_application_circuit_data.into_iter() {
-        circuit_callback(ZkSyncBaseLayerCircuit::StorageApplication(
-            maker.process(circuit_input, circuit_type),
-        ));
-    }
-
-    let (
-        storage_application_circuits,
-        queue_simulator,
-        storage_application_circuits_compact_forms_witnesses,
-    ) = maker.into_results();
-    all_compact_forms.extend(storage_application_circuits_compact_forms_witnesses.clone());
-    queue_simulator_callback(
-        circuit_type as u8,
-        queue_simulator,
-        storage_application_circuits_compact_forms_witnesses,
-    );
-
-    // events sorter
-    let circuit_type = BaseLayerCircuitType::EventsRevertsFilter;
-
-    let mut maker = CircuitMaker::new(
-        geometry.cycles_per_events_or_l1_messages_sorter,
-        round_function.clone(),
-        &mut cs_for_witness_generation,
-        &mut cycles_used,
-    );
-
-    for circuit_input in events_deduplicator_circuit_data.into_iter() {
-        circuit_callback(ZkSyncBaseLayerCircuit::EventsSorter(
-            maker.process(circuit_input, circuit_type),
-        ));
-    }
-
-    let (events_sorter_circuits, queue_simulator, events_sorter_circuits_compact_forms_witnesses) =
-        maker.into_results();
-    all_compact_forms.extend(events_sorter_circuits_compact_forms_witnesses.clone());
-    queue_simulator_callback(
-        circuit_type as u8,
-        queue_simulator,
-        events_sorter_circuits_compact_forms_witnesses,
-    );
-
-    // l1 messages sorter
-    let circuit_type = BaseLayerCircuitType::L1MessagesRevertsFilter;
-
-    let mut maker = CircuitMaker::new(
-        geometry.cycles_per_events_or_l1_messages_sorter,
-        round_function.clone(),
-        &mut cs_for_witness_generation,
-        &mut cycles_used,
-    );
-
-    for circuit_input in l1_messages_deduplicator_circuit_data.into_iter() {
-        circuit_callback(ZkSyncBaseLayerCircuit::L1MessagesSorter(
-            maker.process(circuit_input, circuit_type),
-        ));
-    }
-
-    let (
-        l1_messages_sorter_circuits,
-        queue_simulator,
-        l1_messages_sorter_circuits_compact_forms_witnesses,
-    ) = maker.into_results();
-    all_compact_forms.extend(l1_messages_sorter_circuits_compact_forms_witnesses.clone());
-    queue_simulator_callback(
-        circuit_type as u8,
-        queue_simulator,
-        l1_messages_sorter_circuits_compact_forms_witnesses,
-    );
-
-    // l1 messages pubdata hasher
-    let circuit_type = BaseLayerCircuitType::L1MessagesHasher;
-
-    let mut maker = CircuitMaker::new(
-        geometry.limit_for_l1_messages_pudata_hasher,
-        round_function.clone(),
-        &mut cs_for_witness_generation,
-        &mut cycles_used,
-    );
-
-    for circuit_input in l1_messages_linear_hash_data.into_iter() {
-        circuit_callback(ZkSyncBaseLayerCircuit::L1MessagesHasher(
-            maker.process(circuit_input, circuit_type),
-        ));
-    }
-
-    let (
-        l1_messages_hasher_circuits,
-        queue_simulator,
-        l1_messages_hasher_circuits_compact_forms_witnesses,
-    ) = maker.into_results();
-    all_compact_forms.extend(l1_messages_hasher_circuits_compact_forms_witnesses.clone());
-    queue_simulator_callback(
-        circuit_type as u8,
-        queue_simulator,
-        l1_messages_hasher_circuits_compact_forms_witnesses,
-    );
-
-    // done!
-
-    let basic_circuits = BlockFirstAndLastBasicCircuits {
-        main_vm_circuits,
-        code_decommittments_sorter_circuits,
-        code_decommitter_circuits,
-        log_demux_circuits,
-        keccak_precompile_circuits,
-        sha256_precompile_circuits,
-        ecrecover_precompile_circuits,
-        ram_permutation_circuits,
-        storage_sorter_circuits,
-        storage_application_circuits,
-        events_sorter_circuits,
-        l1_messages_sorter_circuits,
-        l1_messages_hasher_circuits,
-    };
-
-    (basic_circuits, all_compact_forms)
-}
-
 pub struct BlockFirstAndLastBasicCircuits {
     pub main_vm_circuits: FirstAndLastCircuit<
         VmMainInstanceSynthesisFunction<Field, VmWitnessOracle<Field>, RoundFunction>,
@@ -631,7 +142,7 @@ impl<S: ZkSyncUniformSynthesisFunction<Field>> Default for FirstAndLastCircuit<S
 
 /// Implemented for structs that have a field called `closed_form_input`.
 /// They are defined as if they were completely unrelated in era-zkevm_circuits.
-trait ClosedFormInputField<F: SmallField> {
+pub(crate) trait ClosedFormInputField<F: SmallField> {
     type T: Clone
         + std::fmt::Debug
         + CSAllocatable<F>
@@ -791,7 +302,7 @@ impl<F: SmallField> ClosedFormInputField<F> for EventsDeduplicatorInstanceWitnes
     }
 }
 
-struct CircuitMaker<'a, T, S>
+pub(crate) struct CircuitMaker<'a, T, S>
 where
     T: ClosedFormInputField<GoldilocksField>,
     S: ZkSyncUniformSynthesisFunction<
@@ -827,7 +338,7 @@ where
         RoundFunction = Poseidon2Goldilocks,
     >,
 {
-    fn new(
+    pub(crate) fn new(
         geometry: u32,
         round_function: Arc<Poseidon2Goldilocks>,
         cs_for_witness_generation: &'a mut ConstraintSystemImpl<
@@ -848,7 +359,7 @@ where
         }
     }
 
-    fn process(
+    pub(crate) fn process(
         &mut self,
         mut circuit_input: T,
         circuit_type: BaseLayerCircuitType,
@@ -902,7 +413,7 @@ where
         circuit
     }
 
-    fn into_results(
+    pub(crate) fn into_results(
         self,
     ) -> (
         FirstAndLastCircuit<S>,
